refactor(web): extract WS message handler into useChatMessageHandler hook
